Moving CPU info logic into it's own class
